requirements.txt
python-dotenv
PyGithub
google-generativeai
